requirements.txt
Flask==0.11.1
Flask-RESTful==0.3.5
Jinja2==2.8
MarkupSafe==0.23
PyYAML==3.12
Werkzeug==0.11.11
aniso8601==1.2.0
click==6.6
flask-swagger==0.2.13
itsdangerous==0.24
mockito==0.7.0
python-consul==0.6.1
python-dateutil==2.6.0
pytz==2016.7
requests==2.11.1
six==1.10.0
